docs(phase-03): research and plans for CLI phase
Phase 3 CLI: 3 plans in 2 waves
- 03-01: CLI scaffold, readline singleton, dialogue/options UI (Wave 1)
- 03-02: execution stream, violation/escalation handlers, result screen (Wave 2)
- 03-03: benchmark fixtures (10 plans) and runner with 80% threshold (Wave 1)

Also includes Phase 2 artifacts: @treis/errors package extraction,
VERIFICATION.md, and STATE.md updated to phase_complete.
